refactor: MenuGroup 패키지 이동
