+++ src/main/java/kitchenpos/menu/repository/MenuGroupRepository.java
@@ -1,6 +1,6 @@
-package kitchenpos.menugroup.repository;
+package kitchenpos.menu.repository;
 
-import kitchenpos.menugroup.domain.MenuGroup;
+import kitchenpos.menu.domain.MenuGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MenuGroupRepository extends JpaRepository<MenuGroup, Long> {
